+++ djstripe/south_migrations/0006_unique_constraint_invoice_stripe_id.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-import datetime
 from south.db import db
 from south.v2 import SchemaMigration
 
@@ -7,49 +6,14 @@ from south.v2 import SchemaMigration
 class Migration(SchemaMigration):
 
     def forwards(self, orm):
-        # Adding model 'Plan'
-        db.create_table(u'djstripe_plan', (
-            (u'id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
-            ('created', self.gf('model_utils.fields.AutoCreatedField')(default=datetime.datetime.now)),
-            ('modified', self.gf('model_utils.fields.AutoLastModifiedField')(default=datetime.datetime.now)),
-            ('stripe_id', self.gf('django.db.models.fields.CharField')(unique=True, max_length=50)),
-            ('name', self.gf('django.db.models.fields.CharField')(max_length=100)),
-            ('currency', self.gf('django.db.models.fields.CharField')(max_length=10)),
-            ('interval', self.gf('django.db.models.fields.CharField')(max_length=10)),
-            ('interval_count', self.gf('django.db.models.fields.IntegerField')(default=1, null=True)),
-            ('amount', self.gf('django.db.models.fields.DecimalField')(max_digits=7, decimal_places=2)),
-            ('trial_period_days', self.gf('django.db.models.fields.IntegerField')(null=True)),
-        ))
-        db.send_create_signal(u'djstripe', ['Plan'])
+        # Adding unique constraint on 'Invoice', fields ['stripe_id']
+        db.create_unique(u'djstripe_invoice', ['stripe_id'])
 
     def backwards(self, orm):
-        # Deleting model 'Plan'
-        db.delete_table(u'djstripe_plan')
+        # Removing unique constraint on 'Invoice', fields ['stripe_id']
+        db.delete_unique(u'djstripe_invoice', ['stripe_id'])
 
     models = {
-        u'accounts.siuser': {
-            'Meta': {'object_name': 'SIUser'},
-            'date_joined': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
-            'email': ('django.db.models.fields.EmailField', [], {'max_length': '255'}),
-            'first_name': ('django.db.models.fields.CharField', [], {'max_length': '255', 'null': 'True'}),
-            'groups': ('django.db.models.fields.related.ManyToManyField', [], {'to': u"orm['auth.Group']", 'symmetrical': 'False', 'blank': 'True'}),
-            'handle': ('django.db.models.fields.SlugField', [], {'max_length': '255'}),
-            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
-            'is_active': ('django.db.models.fields.BooleanField', [], {'default': 'True'}),
-            'is_admin': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
-            'is_alumni': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
-            'is_applicant': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
-            'is_instructor': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
-            'is_staff': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
-            'is_student': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
-            'is_superuser': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
-            'last_login': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
-            'last_modifiction': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
-            'last_name': ('django.db.models.fields.CharField', [], {'max_length': '255', 'null': 'True'}),
-            'password': ('django.db.models.fields.CharField', [], {'max_length': '128'}),
-            'user_permissions': ('django.db.models.fields.related.ManyToManyField', [], {'to': u"orm['auth.Permission']", 'symmetrical': 'False', 'blank': 'True'}),
-            'username': ('django.db.models.fields.EmailField', [], {'unique': 'True', 'max_length': '255', 'db_index': 'True'})
-        },
         u'auth.group': {
             'Meta': {'object_name': 'Group'},
             u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
@@ -62,6 +26,22 @@ class Migration(SchemaMigration):
             'content_type': ('django.db.models.fields.related.ForeignKey', [], {'to': u"orm['contenttypes.ContentType']"}),
             u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
             'name': ('django.db.models.fields.CharField', [], {'max_length': '50'})
+        },
+        u'auth.user': {
+            'Meta': {'object_name': 'User'},
+            'date_joined': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
+            'email': ('django.db.models.fields.EmailField', [], {'max_length': '75', 'blank': 'True'}),
+            'first_name': ('django.db.models.fields.CharField', [], {'max_length': '30', 'blank': 'True'}),
+            'groups': ('django.db.models.fields.related.ManyToManyField', [], {'symmetrical': 'False', 'related_name': "u'user_set'", 'blank': 'True', 'to': u"orm['auth.Group']"}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'is_active': ('django.db.models.fields.BooleanField', [], {'default': 'True'}),
+            'is_staff': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'is_superuser': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'last_login': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
+            'last_name': ('django.db.models.fields.CharField', [], {'max_length': '30', 'blank': 'True'}),
+            'password': ('django.db.models.fields.CharField', [], {'max_length': '128'}),
+            'user_permissions': ('django.db.models.fields.related.ManyToManyField', [], {'symmetrical': 'False', 'related_name': "u'user_set'", 'blank': 'True', 'to': u"orm['auth.Permission']"}),
+            'username': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '30'})
         },
         u'contenttypes.contenttype': {
             'Meta': {'ordering': "('name',)", 'unique_together': "(('app_label', 'model'),)", 'object_name': 'ContentType', 'db_table': "'django_content_type'"},
@@ -119,7 +99,7 @@ class Migration(SchemaMigration):
             u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
             'modified': ('model_utils.fields.AutoLastModifiedField', [], {'default': 'datetime.datetime.now'}),
             'stripe_id': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '50'}),
-            'user': ('django.db.models.fields.related.OneToOneField', [], {'to': u"orm['accounts.SIUser']", 'unique': 'True', 'null': 'True'})
+            'subscriber': ('django.db.models.fields.related.OneToOneField', [], {'to': u"orm['auth.User']", 'unique': 'True', 'null': 'True'})
         },
         u'djstripe.event': {
             'Meta': {'object_name': 'Event'},
@@ -133,7 +113,7 @@ class Migration(SchemaMigration):
             'stripe_id': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '50'}),
             'valid': ('django.db.models.fields.NullBooleanField', [], {'null': 'True', 'blank': 'True'}),
             'validated_message': ('jsonfield.fields.JSONField', [], {'null': 'True'}),
-            'webhook_message': ('jsonfield.fields.JSONField', [], {'default': '{}'})
+            'webhook_message': ('jsonfield.fields.JSONField', [], {})
         },
         u'djstripe.eventprocessingexception': {
             'Meta': {'object_name': 'EventProcessingException'},
@@ -159,7 +139,7 @@ class Migration(SchemaMigration):
             'paid': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
             'period_end': ('django.db.models.fields.DateTimeField', [], {}),
             'period_start': ('django.db.models.fields.DateTimeField', [], {}),
-            'stripe_id': ('django.db.models.fields.CharField', [], {'max_length': '50'}),
+            'stripe_id': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '50'}),
             'subtotal': ('django.db.models.fields.DecimalField', [], {'max_digits': '7', 'decimal_places': '2'}),
             'total': ('django.db.models.fields.DecimalField', [], {'max_digits': '7', 'decimal_places': '2'})
         },