+++ HISTORY.rst
@@ -3,11 +3,18 @@
 History
 =======
 
-0.5.0 (2015-04-??)
+0.5.0 (2015-05-??)
 ---------------------
 
 * Removed the StripeSubscriptionSignupForm
+* Removed ``djstripe.safe_settings``. Settings are now all located in ``djstripe.settings``
+* ``DJSTRIPE_TRIAL_PERIOD_FOR_SUBSCRIBER_CALLBACK`` can no longer be a module string
+* The sync_subscriber argument has been renamed from subscriber_model to subscriber
 * Moved available currencies to the DJSTRIPE_CURRENCIES setting (Thanks @martinhill)
+* Allow passing of extra parameters to stripe Charge API (Thanks @mthornhill)
+* charge.refund() now returns the refunded charge object (Thanks @mthornhill)
+* Charge model now has captured field and a capture method (Thanks @mthornhill)
+* South migrations are now up to date (Thanks @Tyrdall)
 
 0.4.0 (2015-04-05)
 ----------------------